% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{lmPredictAge}
\alias{lmPredictAge}
\title{Linear regression age prediction}
\usage{
lmPredictAge(dataset)
}
\arguments{
\item{dataset}{DNA methylation matrix}
}
\value{
Predicted age
}
\description{
Using linear regression to predict age
}
\examples{
predictedAge <- lmPredictAge(beta)
}
