% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{rfPredictAge}
\alias{rfPredictAge}
\title{Random forest age prediction}
\usage{
rfPredictAge(dataset)
}
\arguments{
\item{dataset}{DNA methylation matrix}
}
\value{
Predicted age
}
\description{
Using random forest to predict age
}
\examples{
predictedAge <- rfPredictAge(beta)
}
