% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{svmPredictAge}
\alias{svmPredictAge}
\title{Support vector machines age prediction}
\usage{
svmPredictAge(dataset)
}
\arguments{
\item{dataset}{DNA methylation matrix}
}
\value{
Predicted age
}
\description{
Using support vector machines to predict age
}
\examples{
predictedAge <- svmPredictAge(beta)
}
